GH-470: Add SeekToCurrentErrorHandler
Fixes: https://github.com/spring-projects/spring-kafka/issues/470

2.0 introduced the `ConsumerAwareErrorHandler`, which allows the user to
`seek()` the consumer after an exception.

However, any pending records from the previous poll are still sent to
the listener.

Add the `RemainingRecordsErrorHandler` interface and `SeekToCurrentErrorHandler`
implementation which resets the offsets of all unprocessed topic/partitions so
that the next poll will return the failed record and all of the previously
unprocessed records.

__cherry-pick to 2.0.x__

* Fix concurrent access to mocks - stop containers

* Fix race - wait for commits

* Test polishing - wait for 3rd `poll()`

* One more commit - `onPartitionsAssigned()`

* Polishing - PR Comments

(cherry picked from commit c42df50)